vector/requirements.txt
langchain==0.2.3
chromadb==0.5.1
faiss-cpu==1.8.1
openai==1.30.0
tiktoken==0.6.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pandas==2.2.2
llama-index==0.10.28.post1
sentence-transformers==2.6.1
numpy>=1.26 